requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
requests>=2.25.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
